Move asset download into SetupAssets and add progress logging
